+++ includes/bp-avatar-suggestions-admin.php
@@ -1,6 +1,6 @@
 <?php
 // Exit if accessed directly
-if ( ! defined( 'ABSPATH' ) ) exit;
+defined( 'ABSPATH' ) || exit;
 /**
  * Load Admin class.
  *